Start refactoring KWDataGridOptimizer with VxV and IxV sub-classes
Premiere etape:
- KWDataGridOptimizer: classe ancetre de l'optimisation, essentiellement focalisee sur l'implementation de l'algorithme VNS
- KWDataGridOptimizerVxV: sous classe dediee au cas VxV
  - methodes dediees a la boucle sur la granularite
- KWDataGridOptimizerIxV: sous classe dediee au cas VxV
  - methodes dediees a la boucle sur la tokenisation
- CCCoclusteringBuilder: nettoyage et simplification

Non pris en compte:
- nettoyage de KWDataGridOptimizer, dont les specifilisation dediee VarPart devront etre deplacee vers KWDataGridOptimizerIxV

Details du refactoring
----------------------

Utilisation de l'optimiseur de KWDataGridOptimizerVxV au lieu de KWDataGridOptimizer
- CCCoclusteringBuilder::OptimizeDataGrid
- KWAttributeSubsetStats::ComputeStats

Utilisation de l'optimiseur de KWDataGridOptimizerIxV  au lieu de KWDataGridOptimizer
- CCCoclusteringBuilder::OptimizeVarPartDataGrid

La granularite d'une grille de type VarPart est necessairement 0
- KWDataGrid::Check

La methode KWDataGridOptimizer::OptimizeDataGrid devient virtuelle,
et redefinie de facon specifique dans les sous-classes VxV et IxV

Deplacement de methodes dediees de KWDataGridOptimizer vers KWDataGridOptimizerVxV
- methodes dediee a la granularite
  - IsLastGranularity
  - ComputeMaxExploredGranularity
  - OptimizeGranularizedDataGrid
  - SlightOptimizeGranularizedDataGrid
  - PostOptimizeGranularity
- initialisation dediees au cas supervise
  - OptimizeWithBestUnivariatePartitionForCurrentGranularity
  - OptimizeWithMultipleUnivariatePartitions
- methode d'affichage specialisee
  - DisplayProgression

Principaux autre impacts
- KWDataGridOptimizer
  - OptimizeDataGrid: gestion commune de l'optimisation, redirigee ver InternalOptimizeDataGrid
    - donnees de travail initialisee et accessibles dans toutes les methodes d'optimisation
      - GetOptimizedInitialDataGrid: grille initiale
      - GetOptimizedNullDataGrid: model null
      - GetOptimizedNullDataGridCost: cout du model null
  - InternalOptimizeDataGrid: optimisation specifique, reimplementee dans les sous-classes VxV et IxV
  - InitializeWithTerminalDataGrid: suppression, car desormais remplace par une coppie du model null
  - Set|GetInitialVarPartDataGrid: suppression, car desormais remplace par l'acces a la grille initiale
  - Set|GetAttributeSubsetStats -> Set|GetOptimizationHandler
  - VNSOptimizeVarPartDataGrid: supression du dernier parametre, pour unification avec VNSOptimizeDataGrid
- CCCoclusteringBuilder
  - OptimizeDataGrid: appel directement de dataGridOptimizerIxV apres son parametrage
  - OptimizeVarPartDataGrid, PROTO_OptimizeVarPartDataGrid: deplacement vers KWDataGridOptimizerIxV
  - HandleOptimizationStep: calcul de ComputeHierarchicalInfo avec la grille initiale dans le cas VarPart
  - bIsDefaultCostComputed, dAnyTimeDefaultCost: supression, car on utilise systematiquement de cout du modele null donne par des dataGridCosts
  - InitializeDataGridOptimizer: initialisation du dAnyTimeBestCost avec le cout du modele null
- KWDataGridOptimizerIxV
  - OptimizeDataGrid: appel de OptimizeVarPartDataGrid
  - OptimizeVarPartDataGrid, PROTO_OptimizeVarPartDataGrid: adaptation des methodes provenant de KWDataGridOptimizerIxV
- KWDataGridManager::ExportTerminalDataGrid: supression
  - utilisation uniquement de ExportNullDataGrid, dans
    - KWDataGridCosts::InitializeDefaultCosts
    - KWDataGridOptimizer::InitializeWithTerminalDataGrid
  - ExportDataGridWithRandomizedInnerAttributes: supression du parametre referenceInnerAttributes, redondant avec celui de inputDataGrid
  - ExportDataGridWithMergedInnerAttributes: supression du parametre inutile sourceDataGrid
  - ExportReferenceDataGridWithGivenInnerAttributes: supression de la methode inutilisee

Autres nettoyages de detail:
- Profiler::GetMethodStartNumber: code retour de int a longint
- nettoyage des friend class inutiles
  - DTGrouperMODL
  - DTDiscretizerMODL
