+++ src/Learning/KWDataPreparation/KWDataGridOptimizer.cpp
@@ -11,19 +11,23 @@ KWDataGridOptimizer::KWDataGridOptimizer()
 {
 	classStats = NULL;
 	dataGridCosts = NULL;
-	initialVarPartDataGrid = NULL;
-	attributeSubsetStatsHandler = NULL;
+	attributeSubsetStatsOptimizationHandler = NULL;
 	dEpsilon = 1e-6;
+	optimizedInitialDataGrid = NULL;
+	optimizedNullDataGrid = NULL;
 	ResetProgressionIndicators();
 }
 
-KWDataGridOptimizer::~KWDataGridOptimizer() {}
+KWDataGridOptimizer::~KWDataGridOptimizer()
+{
+	assert(optimizedInitialDataGrid == NULL);
+	assert(optimizedNullDataGrid == NULL);
+}
 
 void KWDataGridOptimizer::Reset()
 {
 	classStats = NULL;
 	dataGridCosts = NULL;
-	initialVarPartDataGrid = NULL;
 	ResetProgressionIndicators();
 }
 
@@ -52,48 +56,14 @@ KWClassStats* KWDataGridOptimizer::GetClassStats() const
 	return classStats;
 }
 
-void KWDataGridOptimizer::SetInitialVarPartDataGrid(KWDataGrid* refDataGrid)
-{
-	require(refDataGrid != NULL);
-	require(refDataGrid->IsVarPartDataGrid());
-	initialVarPartDataGrid = refDataGrid;
-}
-
-KWDataGrid* KWDataGridOptimizer::GetInitialVarPartDataGrid() const
-{
-	return initialVarPartDataGrid;
-}
-
 double KWDataGridOptimizer::OptimizeDataGrid(const KWDataGrid* initialDataGrid, KWDataGrid* optimizedDataGrid) const
 
 {
-	boolean bDisplayResults = false;
-	boolean bDisplayGranularities = false;
-	KWDataGrid granularizedDataGrid;
-	KWDataGrid* granularizedOptimizedDataGrid;
 	KWDataGridManager dataGridManager;
-	double dGranularityBestCost;
 	double dBestCost;
-	boolean bIsOptimizationNeeded;
-	ObjectDictionary odQuantileBuilders;
-	IntVector ivMaxPartNumbers;
-	IntVector ivPreviousPartNumber;
-	IntVector ivCurrentPartNumber;
-	const double dMinimalGranularityIncreasingCoefficient = 2;
-	int nMaxExploredGranularity;
-	int nGranularityIndex;
-	boolean bIsLastGranularity;
-	boolean bIsGranularitySelected;
-	int nAttribute;
-	double dBestMergedCost;
-	KWDataGrid granularizedPostMergedOptimizedDataGrid;
-	KWDataGrid partitionedReferenceGranularizedPostMergedDataGrid;
-	int nCurrentExploredGranularity;
-	int nLastExploredGranularity;
-	ALString sSuffix;
-	ALString sTmp;
-	KWDataGrid nullDataGrid;
 
+	require(GetDataGridCosts() != NULL);
+	require(GetDataGridCosts()->IsInitialized());
 	require(initialDataGrid != NULL);
 	require(optimizedDataGrid != NULL);
 
@@ -105,220 +75,29 @@ double KWDataGridOptimizer::OptimizeDataGrid(const KWDataGrid* initialDataGrid, 
 	DisplayOptimizationHeaderLine();
 
 	//Initialisations
-	dGranularityBestCost = DBL_MAX;
-	dBestMergedCost = dGranularityBestCost;
 	ResetProgressionIndicators();
 	timerOptimization.Start();
+	optimizedInitialDataGrid = initialDataGrid;
+
+	// Initialisation du modele null
+	optimizedNullDataGrid = new KWDataGrid;
+	dataGridManager.ExportNullDataGrid(initialDataGrid, optimizedNullDataGrid);
+	assert(GetOptimizedNullDataGridCost() == GetDataGridCosts()->ComputeDataGridTotalCost(optimizedNullDataGrid));
 
 	// Controle de la graine aleatoire pour avoir des resultats reproductibles
 	SetRandomSeed(1);
 
-	// Construction d'une grille terminale pour la solution initiale
-	dBestCost = InitializeWithTerminalDataGrid(initialDataGrid, optimizedDataGrid);
-	if (bDisplayResults)
-	{
-		cout << "OptimizeDataGrid: Cout grille terminale independant de la granularite " << dBestCost << endl;
-		cout << "OptimizeDataGrid: Grille initiale avant optimisation" << endl;
-		initialDataGrid->Write(cout);
-	}
+	// Utilisation du modele nul pour la solution initiale
+	SaveDataGrid(GetOptimizedNullDataGrid(), optimizedDataGrid);
+	dBestCost = GetOptimizedNullDataGridCost();
 
-	// On determine si on peut potentiellement faire mieux que la grille terminale
-	bIsOptimizationNeeded = IsOptimizationNeeded(initialDataGrid);
+	// Optimisation a partir d'une grille initiale complete si algorithme glouton
+	if (not TaskProgression::IsInterruptionRequested())
+		dBestCost = InternalOptimizeDataGrid(initialDataGrid, optimizedDataGrid);
 
-	// Cas ou la grille terminale est ameliorable
-	if (bIsOptimizationNeeded and not TaskProgression::IsInterruptionRequested())
-	{
-		// Calcul de la granularite max a explorer
-		nMaxExploredGranularity = ComputeMaxExploredGranularity(initialDataGrid);
-
-		// Initialisation des quantiles builders a partir de la grille source
-		dataGridManager.InitializeQuantileBuilders(initialDataGrid, &odQuantileBuilders, &ivMaxPartNumbers);
-		if (bDisplayResults)
-			cout << "ivMaxPartNumbers Granularisation\t" << ivMaxPartNumbers << flush;
-
-		// Initialisation des vecteurs de nombre de parties courant et precedent
-		ivPreviousPartNumber.SetSize(ivMaxPartNumbers.GetSize());
-		ivCurrentPartNumber.SetSize(ivMaxPartNumbers.GetSize());
-
-		// Initialisation
-		nCurrentExploredGranularity = -1;
-		nLastExploredGranularity = -1;
-
-		// Parcours des granularites
-		nGranularityIndex = 1;
-		// Dans le cas d'une grille VarPart, on teste uniquement de la granularite max
-		// Cela eviter les attributVarPart avec fourre-tout composes de PV d'innerAttribute differents en VarPartSet initiaux
-		if (initialDataGrid->IsVarPartDataGrid())
-			nGranularityIndex = nMaxExploredGranularity;
-		bIsLastGranularity = false;
-		while (nGranularityIndex <= nMaxExploredGranularity and not bIsLastGranularity)
-		{
-			// Arret si interruption utilisateur
-			if (TaskProgression::IsInterruptionRequested())
-				break;
-
-			// Granularisation de la grille initiale
-			dataGridManager.ExportGranularizedDataGrid(initialDataGrid, &granularizedDataGrid,
-								   nGranularityIndex, &odQuantileBuilders);
-
-			//////////////////////////////////////////////////////////////////////////////////////////////
-			// On determine si la granularite courante doit etre traitee
-			// - bIsGranularitySelected: parce qu'elle differe suffisamment de la granularite precedente
-			//   et de la granularite max
-			// - bIsLastGranularity: par ce que c'est la derniere
-
-			// Calcul du nombre de nombre de parties par attribut attribut pour la granularite en cours
-			for (nAttribute = 0; nAttribute < granularizedDataGrid.GetAttributeNumber(); nAttribute++)
-			{
-				// Memorisation du nombre de parties de l'attribut granularise
-				ivCurrentPartNumber.SetAt(
-				    nAttribute, cast(KWDGAttribute*, granularizedDataGrid.GetAttributeAt(nAttribute))
-						    ->GetPartNumber());
-			}
-
-			// On determine si on a atteint la granularite max
-			bIsLastGranularity = true;
-			if (nGranularityIndex < nMaxExploredGranularity)
-			{
-				for (nAttribute = 0; nAttribute < granularizedDataGrid.GetAttributeNumber();
-				     nAttribute++)
-				{
-					// La granularite n'est pas max si le nombre de parties courant est inferieur
-					// au nombre max de parties pour au moins un attribut
-					if (ivCurrentPartNumber.GetAt(nAttribute) < ivMaxPartNumbers.GetAt(nAttribute))
-					{
-						bIsLastGranularity = false;
-						break;
-					}
-				}
-			}
-
-			// Dans le cas de la granularite max, on positionne l'index de granularite au maximum
-			// pour des raisons d'affichage (pour ne pas afficher la granularite courante)
-			if (bIsLastGranularity)
-				granularizedDataGrid.SetGranularity(nMaxExploredGranularity);
-			assert(bIsLastGranularity == IsLastGranularity(&granularizedDataGrid));
-
-			// Test si on doit traiter la granularite
-			bIsGranularitySelected = false;
-			if (IsOptimizationNeeded(&granularizedDataGrid))
-			{
-				// Dans le cas d'une grille avec un attribut instances et un attribut VarPart, tant que
-				// l'attribut instances ne contient qu'une seule partie, la granularite n'est pas selectionnee.
-				// La granularisation de l'attribut instances est reduite au fourre-tout (1 seule partie)
-				// tant que le nombre d'observations par instances n'est pas superieur a l'effectif minimal N/2^G
-				// Pour un nombre d'observations egal au nombre de variables pour toutes les instances,
-				// il faut atteindre G tel que G > Gmax - log(K) / log(2)
-				//
-				// On ne traite cette granularite que si elle est differe suffisamment de la precedente et de la
-				for (nAttribute = 0; nAttribute < granularizedDataGrid.GetAttributeNumber();
-				     nAttribute++)
-				{
-					// Cas d'accroissement suffisant du nombre de parties
-					if ((ivCurrentPartNumber.GetAt(nAttribute) >=
-					     ivPreviousPartNumber.GetAt(nAttribute) *
-						 dMinimalGranularityIncreasingCoefficient) and
-					    (ivCurrentPartNumber.GetAt(nAttribute) *
-						 dMinimalGranularityIncreasingCoefficient <=
-					     ivMaxPartNumbers.GetAt(nAttribute)))
-					{
-						bIsGranularitySelected = true;
-						break;
-					}
-				}
-			}
-
-			// Affichage de l'info, sur la granularite
-			if (bDisplayGranularities)
-			{
-				cout << "OptimizeDataGrid: granularity " << nGranularityIndex << ": "
-				     << BooleanToString(bIsGranularitySelected);
-				if (bIsLastGranularity)
-					cout << " (last)";
-				cout << endl;
-			}
-
-			// Exploration de la granularite courante
-			if (IsOptimizationNeeded(&granularizedDataGrid) and
-			    (bIsGranularitySelected or bIsLastGranularity))
-			{
-				assert(bIsGranularitySelected or bIsLastGranularity);
-
-				// Memorisation des granularites exploitees
-				nLastExploredGranularity = nCurrentExploredGranularity;
-				nCurrentExploredGranularity = nGranularityIndex;
-
-				// Initialisation de la grille granularisee optimisee a la grille terminale
-				granularizedOptimizedDataGrid = new KWDataGrid;
-				dGranularityBestCost = InitializeWithTerminalDataGrid(&granularizedDataGrid,
-										      granularizedOptimizedDataGrid);
-				if (bDisplayResults)
-				{
-					cout
-					    << "OptimizeDataGrid: Cout Grille initiale granularisee pour granularite = "
-					    << IntToString(nGranularityIndex) << "\t"
-					    << dataGridCosts->ComputeDataGridTotalCost(&granularizedDataGrid) << endl;
-					granularizedDataGrid.Write(cout);
-					granularizedDataGrid.WriteAttributes(cout);
-					granularizedDataGrid.WriteAttributeParts(cout);
-					cout << "OptimizeDataGrid: Cout grille terminale pour cette granularite\t"
-					     << dGranularityBestCost << endl;
-				}
-
-				// Arret si interruption utilisateur
-				if (TaskProgression::IsInterruptionRequested())
-					break;
-
-				// Parametrage du profiling
-				sSuffix = "";
-				if (bIsLastGranularity)
-					sSuffix = " (last)";
-				KWDataGridOptimizer::GetProfiler()->BeginMethod("Optimize granularity" + sSuffix);
-				KWDataGridOptimizer::GetProfiler()->WriteKeyInt("Granularity index", nGranularityIndex);
-				if (optimizedDataGrid->IsVarPartDataGrid())
-					KWDataGridOptimizer::GetProfiler()->WriteKeyInt(
-					    "VarPart granularity",
-					    optimizedDataGrid->GetInnerAttributes()->GetVarPartGranularity());
-
-				if (bDisplayGranularities)
-				{
-					cout << "OptimizeDataGrid\tGranularite\t" << nGranularityIndex << "\n";
-					if (granularizedDataGrid.IsVarPartDataGrid())
-						cout << "attribut VarPart granularise : initial value number\t"
-						     << granularizedDataGrid.GetVarPartAttribute()
-							    ->GetInitialValueNumber()
-						     << endl;
-				}
-
-				// Optimisation de la grille granularisee
-				// Cas non supervise pour les granularites intermediaires: optimisation legere
-				if (not IsSupervisedDataGrid(initialDataGrid) and not bIsLastGranularity)
-					dGranularityBestCost = SlightOptimizeGranularizedDataGrid(
-					    &granularizedDataGrid, granularizedOptimizedDataGrid);
-				// Cas supervise ou derniere granularite en non supervise: optimisation profonde
-				else
-					dGranularityBestCost = OptimizeGranularizedDataGrid(
-					    &granularizedDataGrid, granularizedOptimizedDataGrid);
-
-				// Fin du parametrage du profiling
-				KWDataGridOptimizer::GetProfiler()->EndMethod("Optimize granularity" + sSuffix);
-				if (bDisplayResults)
-				{
-					cout << "OptimizeDataGrid: Apres OptimizeGranularizedDataGrid pour Granularite "
-					     << nGranularityIndex << "\t Cout " << dGranularityBestCost << endl;
-					granularizedOptimizedDataGrid->Write(cout);
-				}
-
-				// Dans le cas d'un coclustering instances * variables, le cout obtenu
-				// dGranularityBestCost est le cout de l'antecedent de la meilleure grille avant fusion
-				// des PV d'un meme cluster L'amelioration de cout doit etre mesuree par rapport au cout
-				// de la grille post-fusionnee de granularizedOptimizedDataGrid
-				if (granularizedOptimizedDataGrid->IsVarPartDataGrid())
-				{
-					// Cas d'amelioration du cout
-					if (dGranularityBestCost < dBestCost - dEpsilon)
-					{
-						dBestCost = dGranularityBestCost;
+	// Tri des parties par attribut, pour preparer les affichages de resultats
+	// ainsi que les resultats de preparation des donnees
+	optimizedDataGrid->SortAttributeParts();
 
 						// Memorisation du nouvel optimum
 						dataGridManager.CopyDataGrid(granularizedOptimizedDataGrid,
@@ -355,13 +134,9 @@ double KWDataGridOptimizer::OptimizeDataGrid(const KWDataGrid* initialDataGrid, 
 							     << endl;
 
 						if (granularizedPostMergedOptimizedDataGrid
-							.GetInformativeAttributeNumber() < 2)
-						{
-							dataGridManager.ExportNullDataGrid(initialDataGrid,
-											   &nullDataGrid);
-							HandleOptimizationStep(&nullDataGrid, &granularizedDataGrid,
+							.GetInformativeAttributeNumber() == 0)
+							HandleOptimizationStep(optimizedDataGrid, &granularizedDataGrid,
 									       true);
-						}
 
 						else
 						{
@@ -410,16 +185,7 @@ double KWDataGridOptimizer::OptimizeDataGrid(const KWDataGrid* initialDataGrid, 
 								"coclustering pour la derniere granularite "
 							     << endl;
 
-						if (optimizedDataGrid->GetInformativeAttributeNumber() < 2)
-						{
-							dataGridManager.ExportNullDataGrid(initialDataGrid,
-											   &nullDataGrid);
-							HandleOptimizationStep(&nullDataGrid, &granularizedDataGrid,
-									       true);
-						}
-						else
-							HandleOptimizationStep(optimizedDataGrid, &granularizedDataGrid,
-									       true);
+						HandleOptimizationStep(optimizedDataGrid, &granularizedDataGrid, true);
 					}
 				}
 
@@ -476,6 +242,9 @@ double KWDataGridOptimizer::OptimizeDataGrid(const KWDataGrid* initialDataGrid, 
 
 	// Nettoyage
 	ResetProgressionIndicators();
+	optimizedInitialDataGrid = NULL;
+	delete optimizedNullDataGrid;
+	optimizedNullDataGrid = NULL;
 
 	ensure(optimizedDataGrid->AreAttributePartsSorted() or TaskProgression::IsInterruptionRequested());
 	ensure(fabs(dBestCost - GetDataGridCosts()->ComputeDataGridTotalCost(optimizedDataGrid)) < dEpsilon);
@@ -523,633 +292,53 @@ void KWDataGridOptimizer::HandleOptimizationStep(const KWDataGrid* optimizedData
 						 boolean bIsLastSaving) const
 {
 	// Integration de la granularite
-	if (attributeSubsetStatsHandler != NULL)
+	if (attributeSubsetStatsOptimizationHandler != NULL)
 	{
 		if (optimizedDataGrid->IsVarPartDataGrid())
 		{
 			KWDataGridManager dataGridManager;
 			KWDataGrid initialFromOptimizedDataGrid;
 
-			dataGridManager.ExportDataGridWithMergedInnerAttributes(
-			    GetInitialVarPartDataGrid(), GetInitialVarPartDataGrid(),
-			    optimizedDataGrid->GetInnerAttributes(), &initialFromOptimizedDataGrid);
-			attributeSubsetStatsHandler->HandleOptimizationStep(
+			dataGridManager.ExportDataGridWithMergedInnerAttributes(GetOptimizedInitialDataGrid(),
+										optimizedDataGrid->GetInnerAttributes(),
+										&initialFromOptimizedDataGrid);
+			attributeSubsetStatsOptimizationHandler->HandleOptimizationStep(
 			    optimizedDataGrid, &initialFromOptimizedDataGrid, bIsLastSaving);
 		}
 		else
-			attributeSubsetStatsHandler->HandleOptimizationStep(optimizedDataGrid,
-									    initialGranularizedDataGrid, bIsLastSaving);
+			attributeSubsetStatsOptimizationHandler->HandleOptimizationStep(
+			    optimizedDataGrid, initialGranularizedDataGrid, bIsLastSaving);
 	}
 }
 
-void KWDataGridOptimizer::SetAttributeSubsetStats(const KWAttributeSubsetStats* attributeSubsetStats)
+void KWDataGridOptimizer::SetOptimizationHandler(const KWAttributeSubsetStats* attributeSubsetStats)
 {
-	attributeSubsetStatsHandler = attributeSubsetStats;
+	attributeSubsetStatsOptimizationHandler = attributeSubsetStats;
 }
 
-const KWAttributeSubsetStats* KWDataGridOptimizer::GetAttributeSubsetStats()
+const KWAttributeSubsetStats* KWDataGridOptimizer::GetOptimizationHandler()
 {
-	return attributeSubsetStatsHandler;
+	return attributeSubsetStatsOptimizationHandler;
 }
 
-double KWDataGridOptimizer::OptimizeGranularizedDataGrid(const KWDataGrid* initialDataGrid,
-							 KWDataGrid* optimizedDataGrid) const
+const KWDataGrid* KWDataGridOptimizer::GetOptimizedInitialDataGrid() const
 {
-	boolean bDisplayResults = false;
-	double dBestCost;
-	ALString sTmp;
-
-	require(initialDataGrid != NULL);
-	require(initialDataGrid->Check());
-	require(IsOptimizationNeeded(initialDataGrid));
-	require(optimizedDataGrid != NULL);
-
-	// Initialisation du meilleur cout
-	dBestCost = dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid);
-	if (bDisplayResults)
-	{
-		cout << "Debut OptimizeGranularizedDataGrid " << endl;
-		cout << "Grille initiale" << endl;
-		initialDataGrid->Write(cout);
-		cout << "Grille optimale" << endl;
-		optimizedDataGrid->Write(cout);
-		cout << "Granularite courante \t " << initialDataGrid->GetGranularity() << endl;
-		cout << " Cout grille optimale a la granularite courante " << dBestCost << endl;
-	}
-
-	// Optimisation univariee
-	if (optimizationParameters.GetUnivariateInitialization() and
-	    initialDataGrid->GetTargetValueNumber() > 0
-	    // on ne fait pas d'univarie si un seul attribut
-	    and initialDataGrid->GetAttributeNumber() > 1)
-	{
-		// Recherche d'une amelioration univariee avec prise en compte de la granularite
-		if (not TaskProgression::IsInterruptionRequested())
-			dBestCost = OptimizeWithBestUnivariatePartitionForCurrentGranularity(initialDataGrid,
-											     optimizedDataGrid);
-
-		// Recherche d'une amelioration par croisement des partitions univariees
-		// Integre la granularite
-		if (not TaskProgression::IsInterruptionRequested())
-			dBestCost = OptimizeWithMultipleUnivariatePartitions(initialDataGrid, optimizedDataGrid);
-	}
-
-	// Affichage
-	if (bDisplayResults)
-	{
-		cout << "Grille optimisee avant VNS" << endl;
-		optimizedDataGrid->Write(cout);
-	}
-
-	// Optimisation a partir d'une grille initiale complete si algorithme glouton
-	if (not TaskProgression::IsInterruptionRequested())
-		dBestCost = IterativeVNSOptimizeDataGrid(initialDataGrid, optimizedDataGrid);
-
-	// Tri des parties par attribut, pour preparer les affichages de resultats
-	// ainsi que les resultats de preparation des donnees
-	optimizedDataGrid->SortAttributeParts();
-
-	// Affichage de la grille finale avec ses couts
-	if (bDisplayResults)
-	{
-		// Grille optimisee
-		cout << "Optimized grid\n";
-		if (optimizedDataGrid->GetAttributeNumber() == 2)
-			optimizedDataGrid->WriteCrossTableStats(cout, 0);
-		dataGridCosts->WriteDataGridAllCosts(optimizedDataGrid, cout);
-		cout << endl;
-
-		// Grille finale
-		cout << *optimizedDataGrid << endl;
-	}
-	// Grille finale
-	if (bDisplayResults)
-	{
-		cout << "Granularite courante \t " << optimizedDataGrid->GetGranularity() << endl;
-		cout << " Cout meilleure grille " << dBestCost << endl;
-	}
-
-	// Retour du meilleur cout de codage
-	ensure(fabs(dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) - dBestCost) < dEpsilon);
-	return dBestCost;
+	require(optimizedInitialDataGrid != NULL);
+	return optimizedInitialDataGrid;
 }
 
-double KWDataGridOptimizer::SlightOptimizeGranularizedDataGrid(const KWDataGrid* initialDataGrid,
-							       KWDataGrid* optimizedDataGrid) const
+const KWDataGrid* KWDataGridOptimizer::GetOptimizedNullDataGrid() const
 {
-	boolean bDisplayResults = false;
-	KWDataGridMerger neighbourDataGrid;
-	double dBestCost;
-	double dCost;
-	ALString sTmp;
-
-	require(initialDataGrid != NULL);
-	require(initialDataGrid->Check());
-	require(IsOptimizationNeeded(initialDataGrid));
-	require(not IsSupervisedDataGrid(initialDataGrid));
-	require(initialDataGrid->GetGranularity() < ComputeMaxExploredGranularity(initialDataGrid));
-	require(optimizedDataGrid != NULL);
-
-	// Initialisation du meilleur cout
-	dBestCost = dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid);
-	if (bDisplayResults)
-	{
-		cout << "Debut SlightOptimizeGranularizedDataGrid " << endl;
-		cout << "Grille initiale" << endl;
-		initialDataGrid->Write(cout);
-		cout << "Grille optimale" << endl;
-		optimizedDataGrid->Write(cout);
-		cout << "Granularite courante \t " << initialDataGrid->GetGranularity() << endl;
-		cout << " Cout grille optimale a la granularite courante " << dBestCost << endl;
-	}
-
-	// Parametrage des couts
-	neighbourDataGrid.SetDataGridCosts(GetDataGridCosts());
-
-	// Generation d'une solution dans un voisinage de la meilleure solution
-	KWDataGridOptimizer::GetProfiler()->BeginMethod("Generate neighbour solution");
-	KWDataGridOptimizer::GetProfiler()->WriteKeyInt("Neighbourhood size", 1);
-	GenerateNeighbourSolution(initialDataGrid, optimizedDataGrid, 1, &neighbourDataGrid);
-	KWDataGridOptimizer::GetProfiler()->EndMethod("Generate neighbour solution");
-
-	// Optimisation de cette solution
-	KWDataGridOptimizer::GetProfiler()->BeginMethod("Slight optimize solution");
-	dCost = OptimizeSolution(initialDataGrid, &neighbourDataGrid, false);
-	KWDataGridOptimizer::GetProfiler()->EndMethod("Slight optimize solution");
-
-	// Si amelioration: on la memorise
-	if (dCost < dBestCost - dEpsilon)
-	{
-		dBestCost = dCost;
-		SaveDataGrid(&neighbourDataGrid, optimizedDataGrid);
-
-		// Gestion de la meilleure solution
-		HandleOptimizationStep(optimizedDataGrid, initialDataGrid, false);
-	}
-
-	// Tri des parties par attribut, pour preparer les affichages de resultats
-	// ainsi que les resultats de preparation des donnees
-	optimizedDataGrid->SortAttributeParts();
-
-	// Affichage de la grille finale avec ses couts
-	if (bDisplayResults)
-	{
-		// Grille optimisee
-		cout << "Optimized grid\n";
-		if (optimizedDataGrid->GetAttributeNumber() == 2)
-			optimizedDataGrid->WriteCrossTableStats(cout, 0);
-		dataGridCosts->WriteDataGridAllCosts(optimizedDataGrid, cout);
-		cout << endl;
-
-		// Grille finale
-		cout << *optimizedDataGrid << endl;
-	}
-	// Grille finale
-	if (bDisplayResults)
-	{
-		cout << "Granularite courante \t " << optimizedDataGrid->GetGranularity() << endl;
-		cout << " Cout meilleure grille " << dBestCost << endl;
-	}
-
-	// Retour du meilleur cout de codage
-	ensure(fabs(dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) - dBestCost) < dEpsilon);
-	return dBestCost;
+	require(optimizedNullDataGrid != NULL);
+	ensure(optimizedNullDataGrid->IsVarPartDataGrid() == GetOptimizedInitialDataGrid()->IsVarPartDataGrid());
+	return optimizedNullDataGrid;
 }
 
-double KWDataGridOptimizer::PostOptimizeGranularity(const KWDataGrid* initialDataGrid, KWDataGrid* optimizedDataGrid,
-						    const ObjectDictionary* odQuantileBuilders,
-						    int nLastExploredGranularity) const
+double KWDataGridOptimizer::GetOptimizedNullDataGridCost() const
 {
-	debug(double dInitialOptimizedCost);
-	double dPostOptimizedCost;
-	int nCurrentGranularity;
-	int nBestGranularity;
-	int nGranularityPartileNumber;
-	int nGranularisationIntervalIndex;
-	int nPartitionCumulatedFrequency;
-	int nGranularisationCumulatedFrequency;
-	int nAttributeIndex;
-	boolean bIncompatibleGranularity;
-	boolean bLastExploredGranularity;
-	KWDGAttribute* attribute;
-	KWDGAttribute* initialAttribute;
-	KWDGPart* part;
-	KWQuantileGroupBuilder* quantileGroupBuilder;
-	KWQuantileIntervalBuilder* quantileIntervalBuilder;
-	ObjectArray oaInitialParts;
-	int nCatchAllFirstModalityIndex;
-	int nCatchAllModalityIndex;
-	KWDGPart* catchAllPart;
-	IntVector ivPartileNumber;
-	IntVector ivBestPartileNumber;
-
-	require(initialDataGrid != NULL);
-	require(optimizedDataGrid != NULL);
-	require(IsSupervisedDataGrid(initialDataGrid));
-	require(odQuantileBuilders != NULL);
-	require(nLastExploredGranularity > 0);
-	require(optimizedDataGrid->GetGranularity() > nLastExploredGranularity + 1);
-
-	// Initialisation
-	bLastExploredGranularity = false;
-	bIncompatibleGranularity = false;
-	nCurrentGranularity = optimizedDataGrid->GetGranularity() - 1;
-	nBestGranularity = optimizedDataGrid->GetGranularity();
-	debug(dInitialOptimizedCost = GetDataGridCosts()->ComputeDataGridTotalCost(optimizedDataGrid));
-
-	// Initialisation du nombre de partiles par attribut
-	nAttributeIndex = 0;
-	while (nAttributeIndex < optimizedDataGrid->GetAttributeNumber())
-	{
-		// Extraction de l'attribut courant
-		attribute = optimizedDataGrid->GetAttributeAt(nAttributeIndex);
-
-		ivBestPartileNumber.Add(attribute->GetGranularizedValueNumber());
-		nAttributeIndex++;
-	}
-
-	// Boucle descendante sur les granularites jusqu'a rencontrer l'avant derniere granularite exploree ou une
-	// granularite incompatible avec la partition
-	while (not bLastExploredGranularity and not bIncompatibleGranularity)
-	{
-		nGranularityPartileNumber = (int)pow(2.0, nCurrentGranularity);
-
-		// Initialisation de l'attribut etudie
-		nAttributeIndex = 0;
-		ivPartileNumber.SetSize(0);
-
-		// Boucle sur les attributs tant que l'on n'a pas rencontre un attribut dont la granularisation est
-		// incompatible avec la grille
-		while (nAttributeIndex < optimizedDataGrid->GetAttributeNumber() and not bIncompatibleGranularity)
-		{
-			// Extraction de l'attribut courant
-			attribute = optimizedDataGrid->GetAttributeAt(nAttributeIndex);
-
-			// Reinitialisation du tableau des parties
-			oaInitialParts.SetSize(0);
-
-			// Cas ou le nombre de partiles theorique de cette granularite est superieur ou egal a la taille
-			// de la partition de l'attribut courant
-			if (nGranularityPartileNumber >= attribute->GetPartNumber())
-			{
-				// Cas d'un attribut continu
-				if (attribute->GetAttributeType() == KWType::Continuous)
-				{
-					quantileIntervalBuilder =
-					    cast(KWQuantileIntervalBuilder*,
-						 odQuantileBuilders->Lookup(attribute->GetAttributeName()));
-					// Memorisation du nombre de partiles (theorique pour attribut numerique,
-					// effectif pour attribut categoriel)
-					ivPartileNumber.Add(nGranularityPartileNumber);
-					nGranularityPartileNumber =
-					    quantileIntervalBuilder->ComputeQuantiles(nGranularityPartileNumber);
-
-					// Cas ou le nombre reel de cette granularisation est superieur ou egal a la
-					// taille de la partition
-					if (nGranularityPartileNumber >= attribute->GetPartNumber())
-					{
-						part = attribute->GetHeadPart();
-						nPartitionCumulatedFrequency = part->GetPartFrequency();
-
-						// Initialisation
-						nGranularisationIntervalIndex = 0;
-						nGranularisationCumulatedFrequency =
-						    quantileIntervalBuilder->GetIntervalLastInstanceIndexAt(
-							nGranularisationIntervalIndex) +
-						    1;
-
-						// Parcours des intervalles de la partition optimale et des intervalles
-						// de la granularisation tant qu'il y a compatibilite
-						while (not bIncompatibleGranularity and part != NULL)
-						{
-							while (nPartitionCumulatedFrequency >
-							       nGranularisationCumulatedFrequency)
-							{
-								nGranularisationIntervalIndex++;
-								nGranularisationCumulatedFrequency =
-								    quantileIntervalBuilder
-									->GetIntervalLastInstanceIndexAt(
-									    nGranularisationIntervalIndex) +
-								    1;
-							}
-
-							if (nPartitionCumulatedFrequency <
-							    quantileIntervalBuilder->GetIntervalLastInstanceIndexAt(
-								nGranularisationIntervalIndex) +
-								1)
-								bIncompatibleGranularity = true;
-							// Partie suivante
-							attribute->GetNextPart(part);
-							if (part != NULL)
-								nPartitionCumulatedFrequency +=
-								    part->GetPartFrequency();
-						}
-					}
-					else
-						bIncompatibleGranularity = true;
-				}
-				// Cas d'un attribut categoriel
-				else
-				{
-					quantileGroupBuilder =
-					    cast(KWQuantileGroupBuilder*,
-						 odQuantileBuilders->Lookup(attribute->GetAttributeName()));
-					nGranularityPartileNumber =
-					    quantileGroupBuilder->ComputeQuantiles(nGranularityPartileNumber);
-
-					// Memorisation du nombre de partiles (theorique pour attribut numerique,
-					// effectif pour attribut categoriel)
-					ivPartileNumber.Add(nGranularityPartileNumber);
-
-					// Cas ou le nombre reel de cette granularisation est superieur ou egal a la
-					// taille de la partition
-					if (nGranularityPartileNumber >= attribute->GetPartNumber())
-					{
-						attribute->BuildIndexingStructure();
-
-						// Export des parties de l'attribut initial
-						initialAttribute =
-						    initialDataGrid->SearchAttribute(attribute->GetAttributeName());
-						initialAttribute->ExportParts(&oaInitialParts);
-						// Extraction de l'index de la 1ere modalite du fourre-tout. Le
-						// fourre-tout est le dernier groupe du quantileBuilder
-						nCatchAllFirstModalityIndex =
-						    quantileGroupBuilder->GetGroupFirstValueIndexAt(
-							nGranularityPartileNumber - 1);
-
-						// Extraction de la partie a ajouter dans le groupe
-						catchAllPart =
-						    cast(KWDGPart*, oaInitialParts.GetAt(nCatchAllFirstModalityIndex));
-
-						// Parcours des autres modalites du fourre-tout pour verifier qu'elles
-						// sont dans le meme groupe que la premiere modalite du fourre-tout
-						nCatchAllModalityIndex = nCatchAllFirstModalityIndex + 1;
-						while (not bIncompatibleGranularity and
-						       nCatchAllModalityIndex <=
-							   quantileGroupBuilder->GetGroupLastValueIndexAt(
-							       nGranularityPartileNumber - 1))
-						{
-							part = cast(KWDGPart*,
-								    oaInitialParts.GetAt(nCatchAllModalityIndex));
-							if (part != catchAllPart)
-								bIncompatibleGranularity = true;
-							nCatchAllModalityIndex++;
-						}
-						attribute->DeleteIndexingStructure();
-					}
-					else
-						bIncompatibleGranularity = true;
-				}
-			}
-			else
-				bIncompatibleGranularity = true;
-
-			nAttributeIndex++;
-		}
-
-		// Cas d'une granularite compatible : memorisation de la granularite et du nombre de partiles par
-		// attribut de la grille
-		if (not bIncompatibleGranularity)
-		{
-			nBestGranularity = nCurrentGranularity;
-			ivBestPartileNumber.CopyFrom(&ivPartileNumber);
-		}
-
-		// Prochaine granularite
-		nCurrentGranularity--;
-		bLastExploredGranularity = (nCurrentGranularity == nLastExploredGranularity);
-	}
-
-	// Mise a jour de la granularite de la grille et du nombre de partiles pour chaque attribut
-	optimizedDataGrid->SetGranularity(nBestGranularity);
-	nAttributeIndex = 0;
-	while (nAttributeIndex < optimizedDataGrid->GetAttributeNumber())
-	{
-		// Extraction de l'attribut courant
-		attribute = optimizedDataGrid->GetAttributeAt(nAttributeIndex);
-
-		attribute->SetGranularizedValueNumber(ivBestPartileNumber.GetAt(nAttributeIndex));
-		nAttributeIndex++;
-	}
-
-	// On renvoie le cout le grillle post-optimisee
-	dPostOptimizedCost = GetDataGridCosts()->ComputeDataGridTotalCost(optimizedDataGrid);
-	debug(ensure(dPostOptimizedCost <= dInitialOptimizedCost));
-	return dPostOptimizedCost;
-}
-
-double KWDataGridOptimizer::InitializeWithTerminalDataGrid(const KWDataGrid* initialDataGrid,
-							   KWDataGrid* optimizedDataGrid) const
-{
-	double dBestCost;
-	KWDataGridManager dataGridManager;
-	KWDataGrid terminalDataGrid;
-
-	// Construction d'une grille terminale pour la solution initiale
-	dataGridManager.ExportTerminalDataGrid(initialDataGrid, &terminalDataGrid);
-	dBestCost = dataGridCosts->GetTotalDefaultCost();
-
-	// Memorisation de la meilleure solution initiale
-	SaveDataGrid(&terminalDataGrid, optimizedDataGrid);
-
-	// Tri des parties par attribut, pour preparer les affichages de resultats
-	// ainsi que les resultats de preparation des donnees
-	optimizedDataGrid->SortAttributeParts();
-
-	// Affichage du resulat (ici: grilles initiale et optimisee sont confondues)
-	DisplayOptimizationDetails(optimizedDataGrid, false);
-	DisplayOptimizationDetails(optimizedDataGrid, true);
-
-	// Retour du cout
-	ensure(fabs(dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) - dBestCost) < dEpsilon);
-	return dBestCost;
-}
-
-double
-KWDataGridOptimizer::OptimizeWithBestUnivariatePartitionForCurrentGranularity(const KWDataGrid* initialDataGrid,
-									      KWDataGrid* optimizedDataGrid) const
-{
-	boolean bDisplayResults = false;
-	KWAttributeStats* attributeStats;
-	KWDataGridManager dataGridManager;
-	KWDataGrid univariateDataGrid;
-	KWDGAttribute* initialAttribute;
-	KWDGAttribute* targetAttribute;
-	double dBestCost;
-	double dCost;
-	int nAttribute;
-	boolean bEvaluated;
-	boolean bImproved;
-
-	require(initialDataGrid != NULL);
-	require(optimizedDataGrid != NULL);
-	require(initialDataGrid->GetTargetValueNumber() > 0);
-
-	// Initialisations
-	dBestCost = dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid);
-	if (bDisplayResults)
-		cout << "OptimizeWithBestUnivariatePartition: cout initial " << dBestCost << endl;
-
-	// Retour si pas de statistiques univariees disponibles
-	if (classStats == NULL)
-		return dBestCost;
-
-	// Recherche des grilles univariees deduites des statistiques univariees
-	bImproved = false;
-
-	// Parcours des attributs
-	for (nAttribute = 0; nAttribute < initialDataGrid->GetAttributeNumber(); nAttribute++)
-	{
-		initialAttribute = initialDataGrid->GetAttributeAt(nAttribute);
-
-		// Initialisation de la grille univariee (uniquement si plus de une partie dans la partition)
-		univariateDataGrid.DeleteAll();
-		bEvaluated = false;
-
-		// Extraction de la partition stockee dans l'attributeStats
-		attributeStats = classStats->LookupAttributeStats(initialAttribute->GetAttributeName());
-		bEvaluated = attributeStats->GetLevel() > 0;
-
-		// Cas d'un attribut informatif
-		if (bEvaluated)
-		{
-			// Cas ou la granularite de la meilleure partition univarie de l'attribut ne correspond pas
-			// a la granularite courante de la grille
-			if (attributeStats->GetPreparedDataGridStats()->GetGranularity() !=
-			    initialDataGrid->GetGranularity())
-			{
-				dataGridManager.BuildUnivariateDataGridFromGranularizedPartition(
-				    initialDataGrid, &univariateDataGrid, nAttribute, classStats);
-				bEvaluated = univariateDataGrid.GetAttributeAt(0)->GetPartNumber() > 1;
-			}
-
-			// Cas ou la granularite de la meilleure partition univariee de l'attribut correspond
-			// a la granularite courante de la grille
-			else
-			{
-				dataGridManager.BuildUnivariateDataGridFromAttributeStats(
-				    initialDataGrid, &univariateDataGrid, attributeStats);
-
-				// Transfert du parametrage du fourre-tout
-				targetAttribute = univariateDataGrid.GetAttributeAt(0);
-				targetAttribute->InitializeCatchAllValueSet(initialAttribute->GetCatchAllValueSet());
-			}
-		}
-
-		// Evaluation si la grille univariee a ete construite
-		if (bEvaluated)
-		{
-			// Evaluation de la grille
-			dCost = dataGridCosts->ComputeDataGridTotalCost(&univariateDataGrid);
-			if (bDisplayResults)
-				cout << " APRES Cout de la grille univariee " << dCost << endl;
-
-			// Affichage du resulat (ici: grilles initiale et optimisee sont confondues)
-			DisplayOptimizationDetails(&univariateDataGrid, false);
-			DisplayOptimizationDetails(&univariateDataGrid, true);
-
-			// Memorisation de la meilleure solution
-			if (dCost < dBestCost - dEpsilon)
-			{
-				dBestCost = dCost;
-				SaveDataGrid(&univariateDataGrid, optimizedDataGrid);
-				bImproved = true;
-			}
-		}
-	}
-
-	// Retour du cout
-	ensure(fabs(dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) - dBestCost) == 0 or bImproved);
-	ensure(fabs(dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) - dBestCost) < dEpsilon);
-	return dBestCost;
-}
-
-double KWDataGridOptimizer::OptimizeWithMultipleUnivariatePartitions(const KWDataGrid* initialDataGrid,
-								     KWDataGrid* optimizedDataGrid) const
-{
-	boolean bDisplayResults = false;
-	double dBestCost;
-	double dCost;
-	KWDataGridManager dataGridManager;
-	KWDataGridMerger multivariateDataGrid;
-	KWDataGridPostOptimizer dataGridPostOptimizer;
-	boolean bOk;
-
-	require(initialDataGrid != NULL);
-	require(optimizedDataGrid != NULL);
-	require(initialDataGrid->GetTargetValueNumber() > 0);
-
-	// Initialisations
-	dBestCost = dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid);
-	multivariateDataGrid.SetDataGridCosts(dataGridCosts);
-	dataGridPostOptimizer.SetDataGridCosts(dataGridCosts);
-	if (bDisplayResults)
-	{
-		cout << "OptimizeWithMultipleUnivariatePartitions: dBestCost initial " << dBestCost << endl;
-		cout << " Grille initiale " << endl;
-		cout << *optimizedDataGrid;
-	}
-
-	// Retour si pas assez de statistiques univariees disponibles
-	if (classStats == NULL or classStats->GetInformativeAttributeNumber() <= 1)
-		return dBestCost;
-
-	// Construction d'une grille par croisement des partition univariee
-	// Traitement avec calcul des partitions univariees pour cette granularite
-	bOk = dataGridManager.BuildDataGridFromUnivariateProduct(initialDataGrid, &multivariateDataGrid, classStats);
-
-	if (not bOk)
-		return dBestCost;
-
-	// Affichage du cout initial
-	DisplayOptimizationDetails(&multivariateDataGrid, false);
-
-	// Cout initial si aucune optimisation
-	dCost = DBL_MAX;
-	if (not optimizationParameters.GetOptimize() and not optimizationParameters.GetPreOptimize() and
-	    not optimizationParameters.GetPostOptimize())
-		dCost = dataGridCosts->ComputeDataGridTotalCost(&multivariateDataGrid);
-
-	if (bDisplayResults)
-	{
-		dCost = dataGridCosts->ComputeDataGridTotalCost(&multivariateDataGrid);
-		cout << "OptimizeWithMultipleUnivariatePartitions: dBestCost multivarie initial " << dCost << endl;
-		cout << " Grille initiale " << endl;
-		cout << multivariateDataGrid;
-	}
-
-	// Pre-optimisation de la grille
-	if (optimizationParameters.GetPreOptimize())
-		dCost = dataGridPostOptimizer.PostOptimizeDataGrid(initialDataGrid, &multivariateDataGrid, false);
-
-	// Optimisation par fusion des groupes
-	if (optimizationParameters.GetOptimize())
-		dCost = multivariateDataGrid.Merge();
-
-	// Post-optimisation de la grille
-	if (optimizationParameters.GetPostOptimize())
-		dCost = dataGridPostOptimizer.PostOptimizeDataGrid(initialDataGrid, &multivariateDataGrid, true);
-
-	if (bDisplayResults)
-	{
-		cout << "OptimizeWithMultipleUnivariatePartitions: dBestCost multivarie optimise " << dCost << endl;
-		cout << " Grille multivariee optimisee " << endl;
-		cout << multivariateDataGrid;
-	}
-
-	// Affichage du cout final
-	DisplayOptimizationDetails(&multivariateDataGrid, true);
-
-	// Memorisation de la meilleure solution
-	if (dCost < dBestCost - dEpsilon)
-	{
-		dBestCost = dCost;
-		SaveDataGrid(&multivariateDataGrid, optimizedDataGrid);
-	}
-
-	// Retour du cout
-	ensure(fabs(dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) - dBestCost) < dEpsilon);
-	return dBestCost;
+	require(optimizedNullDataGrid != NULL);
+	require(GetDataGridCosts()->IsInitialized());
+	return GetDataGridCosts()->GetTotalDefaultCost();
 }
 
 double KWDataGridOptimizer::IterativeVNSOptimizeDataGrid(const KWDataGrid* initialDataGrid,
@@ -1163,8 +352,6 @@ double KWDataGridOptimizer::IterativeVNSOptimizeDataGrid(const KWDataGrid* initi
 	KWDataGridManager dataGridManager;
 	KWDataGrid currentDataGrid;
 	int nNeighbourhoodLevelNumber;
-	double dMergedCost;
-	double dBestMergedCost;
 
 	require(initialDataGrid != NULL);
 	require(optimizedDataGrid != NULL);
@@ -1177,12 +364,11 @@ double KWDataGridOptimizer::IterativeVNSOptimizeDataGrid(const KWDataGrid* initi
 	// Parametrage d'un niveau d'optimisation anytime si une limite de temps est indiquee
 	// On le fait uniquement pour la derniere granularite, pour que le mode anytime ne
 	// ne reste pas bloque des la premiere granularite intermediaire
-	if (optimizationParameters.GetOptimizationTime() > 0 and IsLastGranularity(initialDataGrid))
+	if (optimizationParameters.GetOptimizationTime() > 0)
 		nMaxLevel = 20;
 
 	// Initialisations
 	dBestCost = dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid);
-	dBestMergedCost = dBestCost;
 
 	// Appel de VNS en augmentant le nombre de voisinnages d'un facteur 2 chaque fois
 	for (nLevel = 0; nLevel < nMaxLevel; nLevel++)
@@ -1218,8 +404,8 @@ double KWDataGridOptimizer::IterativeVNSOptimizeDataGrid(const KWDataGrid* initi
 		{
 			// On distingue le cout dMergedCost de la meilleure grille et le cout dCost de l'antecedent de
 			// la meilleure grille sans fusion des PV adjacentes dans un meme cluster
-			dCost = VNSOptimizeVarPartDataGrid(initialDataGrid, nNeighbourhoodLevelNumber, &currentDataGrid,
-							   dMergedCost);
+			dCost =
+			    VNSOptimizeVarPartDataGrid(initialDataGrid, nNeighbourhoodLevelNumber, &currentDataGrid);
 			if (dCost < dBestCost - dEpsilon)
 			{
 				dBestCost = dCost;
@@ -1340,7 +526,6 @@ double KWDataGridOptimizer::VNSDataGridPostOptimizeVarPart(const KWDataGrid* ini
 	require(initialDataGrid != NULL);
 	require(neighbourDataGrid != NULL);
 	require(neighbourDataGrid->GetDataGridCosts() == dataGridCosts);
-	require(IsLastGranularity(neighbourDataGrid));
 	require(fabs(dNeighbourDataGridCost - dataGridCosts->ComputeDataGridTotalCost(neighbourDataGrid)) < dEpsilon);
 
 	// Initialisation du cout
@@ -1407,7 +592,7 @@ double KWDataGridOptimizer::VNSDataGridPostOptimizeVarPart(const KWDataGrid* ini
 				if (bDisplayDetails)
 				{
 					cout << "\tGrille initiale du DataGridOptimizer\n";
-					cout << *GetInitialVarPartDataGrid();
+					cout << *GetOptimizedInitialDataGrid();
 					cout << "Grille initiale utilisee pour l'export\n";
 					cout << *initialDataGrid;
 				}
@@ -1431,8 +616,8 @@ double KWDataGridOptimizer::VNSDataGridPostOptimizeVarPart(const KWDataGrid* ini
 				// Parametrage par la grille initiale de l'optimiseur
 				assert(mergedDataGrid->Check());
 				dataGridManager.ExportDataGridWithSingletonVarParts(
-				    GetInitialVarPartDataGrid(), mergedDataGrid, partitionedReferencePostMergedDataGrid,
-				    false);
+				    GetOptimizedInitialDataGrid(), mergedDataGrid,
+				    partitionedReferencePostMergedDataGrid, false);
 				ivGroups.SetSize(
 				    partitionedReferencePostMergedDataGrid->GetVarPartAttribute()->GetPartNumber());
 				assert(partitionedReferencePostMergedDataGrid->Check());
@@ -1553,8 +738,7 @@ double KWDataGridOptimizer::VNSDataGridPostOptimizeVarPart(const KWDataGrid* ini
 }
 
 double KWDataGridOptimizer::VNSOptimizeVarPartDataGrid(const KWDataGrid* initialDataGrid, int nNeighbourhoodLevelNumber,
-						       KWDataGrid* optimizedDataGrid,
-						       double& dBestMergedDataGridCost) const
+						       KWDataGrid* optimizedDataGrid) const
 {
 	boolean bDisplayMainSteps = false;
 	double dBestCost;
@@ -1601,15 +785,12 @@ double KWDataGridOptimizer::VNSOptimizeVarPartDataGrid(const KWDataGrid* initial
 		KWDataGridOptimizer::GetProfiler()->BeginMethod("Generate neighbour solution");
 		KWDataGridOptimizer::GetProfiler()->WriteKeyDouble("Neighbourhood size", dNeighbourhoodSize);
 
-		// ajout pour debbuger OptimizeMerge > modele nul
-		GetInitialVarPartDataGrid()->SetGranularity(ComputeMaxExploredGranularity(GetInitialVarPartDataGrid()));
-
 		// Creation d 'une grille initiale avec les memes innerAttributes que l' optimizedDataGrid
 		// On utilise la grille initialle la plus fine pour la retokenisation
 		// DD 461 initialDataGrid utilisee deux fois pour sourceDataGrid et inputDataGrid : a revoir
-		dataGridManager.ExportDataGridWithMergedInnerAttributes(
-		    GetInitialVarPartDataGrid(), GetInitialVarPartDataGrid(), optimizedDataGrid->GetInnerAttributes(),
-		    &initialFromOptimizedDataGrid);
+		dataGridManager.ExportDataGridWithMergedInnerAttributes(GetOptimizedInitialDataGrid(),
+									optimizedDataGrid->GetInnerAttributes(),
+									&initialFromOptimizedDataGrid);
 		assert(initialFromOptimizedDataGrid.Check());
 
 		// Generation d'un voisin
@@ -1636,9 +817,9 @@ double KWDataGridOptimizer::VNSOptimizeVarPartDataGrid(const KWDataGrid* initial
 		// Optimisation de cette solution
 		// On utilise comme grille de reference celle qui a ete construite pour la generation de la grille voisine
 		initialFromOptimizedDataGrid.DeleteAll();
-		dataGridManager.ExportDataGridWithMergedInnerAttributes(
-		    GetInitialVarPartDataGrid(), GetInitialVarPartDataGrid(), neighbourDataGrid.GetInnerAttributes(),
-		    &initialFromOptimizedDataGrid);
+		dataGridManager.ExportDataGridWithMergedInnerAttributes(GetOptimizedInitialDataGrid(),
+									neighbourDataGrid.GetInnerAttributes(),
+									&initialFromOptimizedDataGrid);
 		dCost = OptimizeSolution(&initialFromOptimizedDataGrid, &neighbourDataGrid, true);
 		KWDataGridOptimizer::GetProfiler()->EndMethod("Optimize solution");
 
@@ -1690,7 +871,7 @@ double KWDataGridOptimizer::VNSOptimizeVarPartDataGrid(const KWDataGrid* initial
 			{
 				// Initialisation de la grille source a la grille tokenisee la plus fine
 				dataGridManager.ExportDataGridWithSingletonVarParts(
-				    GetInitialVarPartDataGrid(), &mergedDataGrid,
+				    GetOptimizedInitialDataGrid(), &mergedDataGrid,
 				    &partitionedReferencePostMergedDataGrid, true);
 				HandleOptimizationStep(&mergedDataGrid, &partitionedReferencePostMergedDataGrid, false);
 
@@ -1716,8 +897,7 @@ double KWDataGridOptimizer::VNSOptimizeVarPartDataGrid(const KWDataGrid* initial
 	}
 
 	// Memorisation du meilleur cout parmi les grilles post-fusionnees
-	dBestMergedDataGridCost = dBestMergedCost;
-	dBestCost = dBestMergedDataGridCost;
+	dBestCost = dBestMergedCost;
 	ensure(fabs(dBestCost - dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid)) < dEpsilon);
 	ensure(optimizedDataGrid->Check());
 	return dBestCost;
@@ -1849,8 +1029,8 @@ void KWDataGridOptimizer::GenerateNeighbourSolution(const KWDataGrid* initialDat
 	{
 		cout << "GenerateNeighbourSolution\t" << dNoiseRate << "\n";
 		cout << "\tGetInitialVarPartDataGrid()\t"
-		     << dataGridCosts->ComputeDataGridTotalCost(GetInitialVarPartDataGrid()) << "\t"
-		     << GetInitialVarPartDataGrid()->GetObjectLabel() << "\n";
+		     << dataGridCosts->ComputeDataGridTotalCost(GetOptimizedInitialDataGrid()) << "\t"
+		     << GetOptimizedInitialDataGrid()->GetObjectLabel() << "\n";
 		cout << "\tinitialDataGrid\t" << dataGridCosts->ComputeDataGridTotalCost(initialDataGrid) << "\t"
 		     << initialDataGrid->GetObjectLabel() << "\n";
 		cout << "\toptimizedDataGrid\t" << dataGridCosts->ComputeDataGridTotalCost(optimizedDataGrid) << "\t"
@@ -1863,7 +1043,7 @@ void KWDataGridOptimizer::GenerateNeighbourSolution(const KWDataGrid* initialDat
 	if (initialDataGrid->IsVarPartDataGrid())
 	{
 		// Nombre de tokens de la grille en entree
-		nInitialCurrentTokenNumber = GetInitialVarPartDataGrid()
+		nInitialCurrentTokenNumber = GetOptimizedInitialDataGrid()
 						 ->GetVarPartAttribute()
 						 ->GetInnerAttributes()
 						 ->ComputeTotalInnerAttributeVarParts();
@@ -1879,8 +1059,7 @@ void KWDataGridOptimizer::GenerateNeighbourSolution(const KWDataGrid* initialDat
 
 		// Generation de la grille surtokenisee
 		dataGridManager.ExportDataGridWithRandomizedInnerAttributes(
-		    GetInitialVarPartDataGrid(), optimizedDataGrid, GetInitialVarPartDataGrid()->GetInnerAttributes(),
-		    &surtokenizedDataGrid, nTargetTokenNumber);
+		    GetOptimizedInitialDataGrid(), optimizedDataGrid, &surtokenizedDataGrid, nTargetTokenNumber);
 		if (bDisplayResults)
 		{
 			cout << "\tTargetTokenNumber\t" << nTargetTokenNumber << "\n";
@@ -1892,9 +1071,9 @@ void KWDataGridOptimizer::GenerateNeighbourSolution(const KWDataGrid* initialDat
 		}
 
 		// Export de la grille antecdent de la grille sur-tokenisee
-		dataGridManager.ExportDataGridWithMergedInnerAttributes(
-		    GetInitialVarPartDataGrid(), GetInitialVarPartDataGrid(), surtokenizedDataGrid.GetInnerAttributes(),
-		    &initialFromSurtokenizedDataGrid);
+		dataGridManager.ExportDataGridWithMergedInnerAttributes(GetOptimizedInitialDataGrid(),
+									surtokenizedDataGrid.GetInnerAttributes(),
+									&initialFromSurtokenizedDataGrid);
 
 		// On remplace le parametrage de la methode par un parametrage dediee au cas VarPart
 		initialDataGrid = &initialFromSurtokenizedDataGrid;
@@ -2053,53 +1232,6 @@ boolean KWDataGridOptimizer::IsSupervisedDataGrid(const KWDataGrid* dataGrid) co
 	return bIsSupervisedDataGrid;
 }
 
-boolean KWDataGridOptimizer::IsLastGranularity(const KWDataGrid* dataGrid) const
-{
-	require(dataGrid != NULL);
-	return dataGrid->GetGranularity() == ComputeMaxExploredGranularity(dataGrid);
-}
-
-int KWDataGridOptimizer::ComputeMaxExploredGranularity(const KWDataGrid* dataGrid) const
-{
-	int nMaxExploredGranularity;
-	const int nMinValueNumber = 500;
-	int nValueNumber;
-	boolean bIsGranularityMaxThresholded;
-
-	require(dataGrid != NULL);
-
-	// Nombre total de valeurs potentielles, base sur le nombre d'individus de la grille
-	nValueNumber = dataGrid->GetGridFrequency();
-
-	// Seuillage eventuel de la granularite max pour reduire la complexite algorithmique,
-	// (sauf dans le cas des petits nombre de valeurs)
-	bIsGranularityMaxThresholded = false;
-	if (nValueNumber > nMinValueNumber)
-	{
-		// Cas d'une grille 2D supervise avec un attribut a predire avec groupage de la cible
-		// (classification avec groupage ou regression) et un attribut explicatif numerique
-		if (dataGrid->GetTargetAttribute() != NULL and dataGrid->GetTargetValueNumber() == 0 and
-		    dataGrid->GetAttributeNumber() == 2 and
-		    dataGrid->GetAttributeAt(0)->GetAttributeType() == KWType::Continuous)
-			bIsGranularityMaxThresholded = true;
-		// Cas d'une classification supervisee simple avec une paire de variables (numerique ou categorielle)
-		else if (dataGrid->GetTargetValueNumber() > 0 and dataGrid->GetAttributeNumber() >= 2 and
-			 nValueNumber > nMinValueNumber)
-			bIsGranularityMaxThresholded = true;
-	}
-
-	// Cas sans seuillage
-	if (not bIsGranularityMaxThresholded)
-		nMaxExploredGranularity = (int)ceil(log(nValueNumber * 1.0) / log(2.0));
-	// Sinon seuillage
-	else
-		nMaxExploredGranularity =
-		    (int)ceil(log(nMinValueNumber + sqrt((nValueNumber - nMinValueNumber) *
-							 log(nValueNumber - nMinValueNumber) / log(2.0))) /
-			      log(2.0));
-	return nMaxExploredGranularity;
-}
-
 boolean KWDataGridOptimizer::IsOptimizationTimeElapsed() const
 {
 	require(timerOptimization.IsStarted());
@@ -2125,7 +1257,6 @@ void KWDataGridOptimizer::DisplayProgression(const KWDataGrid* dataGrid) const
 {
 	double dTotalIterNumber;
 	double dProgression;
-	int nMaxExploredGranularity;
 	ALString sProgressionLabel;
 	double dOptimizationTime;
 	ALString sContext;
@@ -2134,20 +1265,12 @@ void KWDataGridOptimizer::DisplayProgression(const KWDataGrid* dataGrid) const
 	// Calcul du temps d'optimisation
 	dOptimizationTime = timerOptimization.GetElapsedTime();
 
-	// Calcul de la granularite max exploree
-	nMaxExploredGranularity = 0;
-	if (dataGrid->GetGranularity() > 0)
-		nMaxExploredGranularity = ComputeMaxExploredGranularity(dataGrid);
-
 	// Message d'avancement, avec prise en compte de la granularite uniquement si necessaire
 	sProgressionLabel = sTmp + " Iter " + IntToString(nVNSIteration);
 	if (nVNSIteration > 0 and nVNSNeighbourhoodLevelNumber > 0)
 		sProgressionLabel += sTmp + "  VNS " + IntToString(nVNSNeighbourhoodLevelIndex) + "/" +
 				     IntToString(nVNSNeighbourhoodLevelNumber) + " (" +
 				     DoubleToString((int)(10000 * dVNSNeighbourhoodSize) / 10000.0) + ")";
-	if (dataGrid->GetGranularity() > 0 and dataGrid->GetGranularity() < nMaxExploredGranularity)
-		sProgressionLabel += sTmp + "  Granularity " + IntToString(dataGrid->GetGranularity()) + "/" +
-				     IntToString(nMaxExploredGranularity) + " ";
 	TaskProgression::DisplayLabel(sProgressionLabel);
 
 	// Niveau d'avancement avec limite de temps
@@ -2162,17 +1285,8 @@ void KWDataGridOptimizer::DisplayProgression(const KWDataGrid* dataGrid) const
 	// Cas sans limite de temps
 	else
 	{
-		// Calcul du nombre total d'iteration, en combinant granularite et VNS
-		if (IsSupervisedDataGrid(dataGrid))
-		{
-			dTotalIterNumber = pow(2.0, optimizationParameters.GetOptimizationLevel());
-			dTotalIterNumber *= nMaxExploredGranularity;
-		}
-		else
-		{
-			dTotalIterNumber = pow(2.0, optimizationParameters.GetOptimizationLevel());
-			dTotalIterNumber += nMaxExploredGranularity - 1;
-		}
+		// Calcul du nombre total d'iteration
+		dTotalIterNumber = pow(2.0, optimizationParameters.GetOptimizationLevel());
 
 		// Calcul de la progression courante
 		// On tient compte du fait qu'avec l'algorithme VNS, le nombre d'iteration effectif peut depasser
