+++ src/Learning/KWDataPreparation/KWGrouperMODL.h
@@ -83,7 +83,6 @@ public:
 	/////////////////////////////////////////////////////////////////
 	//// Implementation
 protected:
-	friend class KWDataGridOptimizer;
 	friend class KWDataGridManager;
 
 	/////////////////////////////////////////////////////////////////////
